+++ src/charts/index.js
@@ -1,0 +1,3 @@
+import '../globals/chart-styles';
+
+export ColulmnChart from './ColumnChart';