+++ src/globals/chart-styles.js
@@ -1,0 +1,8 @@
+import { injectGlobal } from 'styled-components';
+
+const injectGlobalStyles = () =>
+  injectGlobal`
+    @import 'https://code.highcharts.com/css/highcharts.css';
+  `;
+
+injectGlobalStyles();